Data validation and auth errors

+++ app/Http/Controllers/BooksController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Book;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Validator;
 
 
 class BooksController extends Controller
@@ -11,13 +12,24 @@ class BooksController extends Controller
 
     public function __construct()
     {
-        $this->middleware('auth:api');
+        $this->middleware('auth:api')->except([]);
     }
 
 
     public function store(Request $request){
 
+        $validator = Validator::make($request->all(),[
+            'title' => 'required|max:255',
+            'author' => 'required|max:255',
+            'description' => 'required|max:255'
+        ]);
+
+        if ($validator->fails()) {
+            return response()->json(['errors'=>$validator->errors()]);
+        }
+
         $user_id = auth()->user()->id;
+
         $book = new Book;
         $book->title = $request->input('title');
         $book->author = $request->input('author');
@@ -28,13 +40,29 @@ class BooksController extends Controller
     }
 
     public function index(){
-        $user_id = auth()->user()->id;
-        $books = Book::where('user_id', $user_id)->get();
-        return response()->json(['books' => $books], 200);
+
+        if (auth()->check()) {
+            $user_id = auth()->user()->id;
+            $books = Book::where('user_id', $user_id)->get();
+            return response()->json(['books' => $books], 200);
+        } else {
+            return response()->json(['error' => 'Unauthorized'], 403);
+        }
+
     }
 
     public function destroy($book_id){
+
+        $validator = Validator::make([ 'book_id' => $book_id],[
+            'book_id' => 'required|integer',
+        ]);
+
+        if ($validator->fails()) {
+            return response()->json(['errors'=>$validator->errors()]);
+        }
+
         $user_id = auth()->user()->id;
+
         $book = Book::where('id',$book_id)->where('user_id',$user_id)->first();
         if ($book) {
             $book->delete();
